move renderContext from core to lib, so that every components/containers includes core’s (such as withList, App) can use renderContext.get() to get store, apolloClient, loginToken, history, reducers, middlewares
