+++ packages/nova-core/lib/client/main.js
@@ -1,5 +1,1 @@
-import './auth.js';
-import './mongo_redux.js';
-
 export * from '../modules/index.js';
-export * from './render_context.js';
