+++ packages/nova-lib/lib/client/auth.js
