+++ packages/nova-lib/lib/client/main.js
@@ -1,2 +1,6 @@
+import './auth.js';
+
 export * from '../modules/index.js';
+export * from './mongo_redux.js';
 export * from './apollo.js';
+export * from './render_context.js';
