+++ packages/nova-lib/lib/client/render_context.js
@@ -1,6 +1,7 @@
 import { createStore, applyMiddleware, compose, combineReducers } from 'redux';
 
-import { createApolloClient, addReducer, addMiddleware, getReducers, getMiddlewares } from '../modules/index.js';
+import { addReducer, addMiddleware, getReducers, getMiddlewares } from '../modules/index.js';
+import { createApolloClient } from './apollo.js'
 
 const loginToken = global.localStorage['Meteor.loginToken'];
 const apolloClient = createApolloClient();
