+++ packages/nova-lib/lib/server/main.js
@@ -1,4 +1,8 @@
+import './oauth_config.js';
+import './meteor_subscribe.js'
+
 export * from '../modules/index.js';
 export * from './apollo.js';
 export * from './store.js';
 export * from './mutations.js';
+export * from './render_context.js';
