+++ packages/nova-lib/lib/server/render_context.js
@@ -3,7 +3,9 @@ import { WebApp } from 'meteor/webapp';
 import { DDP } from 'meteor/ddp';
 import { Accounts } from 'meteor/accounts-base';
 
-import { createApolloClient, getReducers, getMiddlewares, configureStore } from '../modules/index.js';
+import { getReducers, getMiddlewares } from '../modules/index.js';
+import { createApolloClient } from './apollo.js';
+import { configureStore } from './store.js';
 
 const Fibers = Npm.require('fibers');
